Clean up pathtracking components and nullpointer deferences
