+++ boomerangPDS/src/main/java/boomerang/pathtracking/DataFlowPathWeight.java
@@ -12,21 +12,25 @@
  *   Johannes Spaeth - initial API and implementation
  * *****************************************************************************
  */
-package boomerang.weights;
+package boomerang.pathtracking;
 
+import boomerang.scope.ControlFlowGraph;
 import boomerang.scope.Statement;
 import boomerang.scope.Val;
 import java.util.Map;
+import java.util.Set;
+import org.jspecify.annotations.NonNull;
+import sync.pds.solver.nodes.Node;
 import wpds.impl.Weight;
 
-public interface PathConditionWeight extends Weight {
-  Map<Statement, ConditionDomain> getConditions();
+public interface DataFlowPathWeight extends Weight {
+  @NonNull PathTrackingWeight getPath();
 
-  Map<Val, ConditionDomain> getEvaluationMap();
+  @NonNull PathConditionWeight getCondition();
 
-  public enum ConditionDomain {
-    TRUE,
-    FALSE,
-    TOP
-  }
+  @NonNull Set<Node<ControlFlowGraph.Edge, Val>> getAllStatements();
+
+  @NonNull Map<Statement, PathConditionWeight.ConditionDomain> getConditions();
+
+  @NonNull Map<Val, PathConditionWeight.ConditionDomain> getEvaluationMap();
 }
