+++ boomerangPDS/src/main/java/boomerang/pathtracking/PathConditionWeightOne.java
@@ -12,7 +12,7 @@
  *   Johannes Spaeth - initial API and implementation
  * *****************************************************************************
  */
-package boomerang.weights;
+package boomerang.pathtracking;
 
 import boomerang.scope.Method;
 import boomerang.scope.Statement;
