+++ boomerangPDS/src/main/java/boomerang/results/QueryResults.java
@@ -15,37 +15,22 @@
 package boomerang.results;
 
 import boomerang.Query;
-import boomerang.scope.Method;
 import java.util.Collection;
-import java.util.Set;
 
 public class QueryResults {
   private final Query query;
-  private final Collection<Method> visitedMethods;
   private final Collection<AffectedLocation> affectedLocations;
-  private final boolean timedout;
 
-  public QueryResults(
-      Query query, Set<AffectedLocation> npes, Set<Method> visMethod, boolean timedout) {
+  public QueryResults(Query query, Collection<AffectedLocation> npes) {
     this.query = query;
-    this.visitedMethods = visMethod;
     this.affectedLocations = npes;
-    this.timedout = timedout;
   }
 
   public Query getQuery() {
     return query;
   }
 
-  public Collection<Method> getVisitedMethods() {
-    return visitedMethods;
-  }
-
   public Collection<AffectedLocation> getAffectedLocations() {
     return affectedLocations;
-  }
-
-  public boolean isTimedout() {
-    return timedout;
   }
 }
